+++ src/domain/usecases/ExportDatasetsUseCase.js
@@ -1,11 +1,12 @@
 export class ExportDatasetsUseCase {
-	constructor(dhisRepository, xlsxRepository) {
-		this.dhisRepository = dhisRepository;
-		this.xlsxRepository = xlsxRepository;
+	constructor(dataSetsDhis2Repository, dataSetsExportSpreadsheetRepository) {
+		this.dataSetsDhis2Repository = dataSetsDhis2Repository;
+		this.dataSetsExportSpreadsheetRepository =
+			dataSetsExportSpreadsheetRepository;
 	}
 
 	execute($resource, dataSetsIds, headers, locales) {
-		return this.dhisRepository
+		return this.dataSetsDhis2Repository
 			.get($resource, dataSetsIds)
 			.$promise.then(({ dataSets }) => {
 				const pickedTranslations = dataSets.map((dataSet) => ({
@@ -25,7 +26,7 @@ export class ExportDatasetsUseCase {
 				const translatedDatasets = pickedTranslations.flatMap(
 					mapDataSetTranslations
 				);
-				const mappedDatasets = this.dhisRepository.getDataSets([
+				const mappedDatasets = getDataSets([
 					...translatedDatasets,
 					...dataSets,
 				]);
@@ -34,7 +35,9 @@ export class ExportDatasetsUseCase {
 					headers: headers.find(({ id }) => id === dataSet.id),
 				}));
 
-				return this.xlsxRepository.createFiles(dataSetsWithHeaders);
+				return this.dataSetsExportSpreadsheetRepository.createFiles(
+					dataSetsWithHeaders
+				);
 			})
 			.then((blobFiles) => {
 				if (blobFiles.length > 1) {
@@ -55,6 +58,114 @@ export class ExportDatasetsUseCase {
 				}
 			});
 	}
+}
+
+function getDataSets(dataSets) {
+	const mappedDatasets = dataSets.flatMap((dataSet) => {
+		if (dataSet.formType === "CUSTOM") return [];
+		const mappedDataSets = {
+			...dataSet,
+			dataSetElements: dataSet.dataSetElements.map(
+				({ dataElement }) => dataElement
+			),
+			sections: dataSet.sections.map(mapSection),
+		};
+
+		return [mappedDataSets];
+	});
+
+	return mappedDatasets;
+}
+
+function mapSection(section) {
+	const mappedCategoryCombos = section.categoryCombos.map((categoryCombo) => {
+		const categories = categoryCombo.categories.map(({ categoryOptions }) =>
+			categoryOptions.map(({ displayFormName }) => displayFormName)
+		);
+
+		const categoryOptionCombos = categoryCombo.categoryOptionCombos
+			.map((categoryOptionCombo) => ({
+				...categoryOptionCombo,
+				categories: categoryOptionCombo.displayFormName.split(", "),
+			}))
+			.filter((categoryOptionCombo) => {
+				const flatCategories = categories.flat();
+				const includesInCategories =
+					categoryOptionCombo.categories.every((category) =>
+						flatCategories.includes(category)
+					);
+				const sameCategoriesLength =
+					categoryOptionCombo.categoryOptions.length ===
+					categoryCombo.categories.length;
+
+				return includesInCategories && sameCategoriesLength;
+			});
+
+		const dataElements = section.dataElements.filter(
+			(de) => de.categoryCombo.id === categoryCombo.id
+		);
+
+		const deIds = dataElements.map(({ id }) => id);
+		const cocIds = categoryOptionCombos.map(({ id }) => id);
+
+		const greyedFields = section.greyedFields.filter(
+			(gf) =>
+				deIds.includes(gf.dataElement.id) &&
+				cocIds.includes(gf.categoryOptionCombo.id)
+		);
+
+		return {
+			...categoryCombo,
+			categoryOptionCombos,
+			categories,
+			dataElements,
+			greyedFields,
+		};
+	});
+
+	//Order category combos by the ones that comes first on the dataset dataElements
+	//Needed when multiple dataElements differ on categoryCombo
+	const orderedCategoryCombos = _.sortBy(
+		mappedCategoryCombos,
+		(categoryCombo) =>
+			section.dataElements.findIndex(
+				(de) => de.categoryCombo.id === categoryCombo.id
+			)
+	);
+
+	const categoryCombos = orderedCategoryCombos.map((categoryCombo) => {
+		return {
+			...categoryCombo,
+			categoryOptionCombos: _.sortBy(
+				categoryCombo.categoryOptionCombos,
+				(categoryOptionCombo) => {
+					//Assign to each word of the (displayFormName) the index where it appears on categoryCombo.categories[]
+					//Output: [1, 2, 0]
+					const prio = categoryOptionCombo.categories.map(
+						(category) => getPrio(categoryCombo, category)
+					);
+
+					//Gives lower priority as [N] increases and does a sum of all values
+					//[1, 2, 0] -> [100, 20, 0] -> 120
+					return prio
+						.map(
+							(v, idx) => v * Math.pow(10, prio.length - 1 - idx)
+						)
+						.reduce((a, b) => a + b, 0);
+				}
+			),
+		};
+	});
+
+	return {
+		...section,
+		categoryCombos,
+	};
+}
+
+function getPrio(categoryCombo, category) {
+	const categories = categoryCombo.categories.flat();
+	return categories.indexOf(category);
 }
 
 function sanitizeFileName(str) {
