+++ src/webapp/controllers/TallySheetsController.js
@@ -28,22 +28,6 @@ export const TallySheetsController = TallySheets.controller(
 			$scope.exportToTable = function (tableId) {
 				const ids = getSelectedDataSets();
 
-				var uri = "data:application/vnd.ms-excel;base64,",
-					template =
-						'<html xmlns:o="urn:schemas-microsoft-com:office:office" ' +
-						'xmlns:x="urn:schemas-microsoft-com:office:excel" xmlns="http://www.w3.org/TR/REC-html40"><head><meta charset="UTF-8">' +
-						"<!--[if gte mso 9]><xml><x:ExcelWorkbook><x:ExcelWorksheets><x:ExcelWorksheet><x:Name>{worksheet}" +
-						"</x:Name><x:WorksheetOptions><x:DisplayGridlines/></x:WorksheetOptions></x:ExcelWorksheet>" +
-						"</x:ExcelWorksheets></x:ExcelWorkbook></xml><![endif]--></head><body><table>{table}</table></body></html>",
-					base64 = function (s) {
-						return window.btoa(unescape(encodeURIComponent(s)));
-					},
-					format = function (s, c) {
-						return s.replace(/{(\w+)}/g, function (m, p) {
-							return c[p];
-						});
-					};
-
 				var table = $("#" + tableId).clone();
 
 				// Remove non-printable section from the table
@@ -58,14 +42,6 @@ export const TallySheetsController = TallySheets.controller(
 
 				// Add border to section table (for printing in MS Excel)
 				table.find(".sectionTable").prop("border", "1");
-
-				// Take the name of the first dataset as filename
-				var name = table.find("h2").first().html() + ".xls";
-
-				var ctx = {
-					worksheet: "MSF-OCBA HMIS" || "Worksheet",
-					table: table.html(),
-				};
 
 				const headers = getHeaders();
 
@@ -125,5 +101,5 @@ function getHeaders() {
 		};
 	});
 
-	return _.orderBy(headers, ({ index }) => index);
+	return _.sortBy(headers, ({ index }) => index);
 }
